Add support for Ultimarc UltraStik360 to act like DS4.
Contains secret combos to switch layouts and allow for joystick merging between joysticks 1 and 2.

+++ ScpControl/ScpDevice.cs
@@ -45,13 +45,13 @@ namespace ScpControl
 
         public string Path { get; protected set; }
 
-        public short VendorId { get; protected set; }
+        public int VendorId { get; protected set; }
 
-        public short ProductId { get; protected set; }
+        public int ProductId { get; protected set; }
 
         protected void GetHardwareId(string devicePath)
         {
-            short vid, pid;
+            int vid, pid;
 
             GetHardwareId(devicePath, out vid, out pid);
 
@@ -60,7 +60,7 @@ namespace ScpControl
             ProductId = pid;
         }
 
-        protected static void GetHardwareId(string devicePath, out short vendorId, out short productId)
+        protected static void GetHardwareId(string devicePath, out int vendorId, out int productId)
         {
             // regex to extract vendor ID and product ID from hardware ID string
             var regex = new Regex("VID_([0-9A-Z]{4})&PID_([0-9A-Z]{4})", RegexOptions.IgnoreCase);
@@ -75,8 +75,8 @@ namespace ScpControl
             }
 
             // get values
-            vendorId = short.Parse(matches[1].Value, NumberStyles.HexNumber);
-            productId = short.Parse(matches[2].Value, NumberStyles.HexNumber);
+            vendorId = int.Parse(matches[1].Value, NumberStyles.HexNumber);
+            productId = int.Parse(matches[2].Value, NumberStyles.HexNumber);
         }
 
         public virtual bool Open(int instance = 0)
